+++ frontend/src/App.js
@@ -2,8 +2,8 @@ import { createContext, useState } from 'react';
 import { BrowserRouter as Router, Redirect, Route, Switch } from 'react-router-dom';
 import * as Realm from 'realm-web';
 import MapVisualiser from './components/map-visualiser/MapVisualiser';
-import LoginPage from './components/sign-in/LoginPage';
-import SignUp from './components/sign-in/SignUp';
+import LoginPage from './components/signing-pages/LoginPage';
+import SignUp from './components/signing-pages/SignUp';
 import Trip from './components/Trip';
 import { useLoadScript } from '@react-google-maps/api';
 import Logo from './components/logo/Logo.jsx';
